% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2sdmean.R
\name{data2sdmean}
\alias{data2sdmean}
\title{Standard deviation and average calculation of a data set.}
\usage{
data2sdmean(x)
}
\arguments{
\item{x}{A list containing 2 or more data.frames of values.}
}
\value{
A list containing the standard deviation and mean values for each variable in x.
}
\description{
This function computes the standard deviation and arithmetic mean values of a data set.
}
\examples{
data=data.setup(list(estr1_A,estr1_B,estr2_A,estr2_B,estr3_A,estr3_B))
data2sdmean(data)
}
