% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.network.A.R
\docType{data}
\name{gene.network.A}
\alias{gene.network.A}
\title{Dataset A of a four-gene network model from Kholodenko et.al.2002}
\format{
A data frame with 4 rows (biological modules) and  5 variables (perturbations):
\describe{
   \item{Modules}{Names of biological modules}
   \item{pert1}{Perturbation targeting gene1}
   \item{pert2}{Perturbation targeting gene2}
   \item{pert3}{Perturbation targeting gene3}
   \item{pert4}{Perturbation targeting gene4}
   \item{basal}{No perturbation}
   ...
}
}
\source{
\url{EDF R\&D}
}
\usage{
gene.network.A
}
\description{
Dataset A of a four-gene network model from Kholodenko et.al.2002
}
\keyword{datasets}
