% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.matrix.R
\name{global.matrix}
\alias{global.matrix}
\title{Global response matrix calculation}
\usage{
global.matrix(data,lb)
}
\arguments{
\item{data}{A data.frame containing experimental data in a specific format.}

\item{lb}{Basal line to calculate the relative activity changes.}
}
\value{
A data frame containing the relative activity changes.
}
\description{
This function calculates the matrix that contains the observed changes in relative activity.
}
\examples{
data=data.setup(list(estr1_A,estr1_B,estr2_A,estr2_B,estr3_A,estr3_B))
Rp=global.matrix(data,lb="E2")
}
