% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{interval}
\alias{interval}
\title{Confidence interval calculation for MRA connectivity coefficients}
\usage{
interval(tab,sd.tab,matp,n=10000,nrep=2)
}
\arguments{
\item{tab}{A data frame containing the experimental data in a specific format.}

\item{sd.tab}{A data frame containing the standard deviation values for each biological module.}

\item{matp}{A perturbation matrix with rows corresponding to biological modules and columns to perturbations.}

\item{n}{Number of iterations for the bootstrap algorithm.}

\item{nrep}{Number of experimental replicas. Default is 2.}
}
\value{
A list containing the upper and lower values of the confidence interval for each connectivity coefficient.
}
\description{
A function that takes in count variability on experimental data replicas and uses the bootstrap algorithm to calculate
            a confidence interval for each MRA connectivity coefficient in a biological network.
}
\examples{
data=data.setup(list(estr1_A,estr1_B,estr2_A,estr2_B,estr3_A,estr3_B))
sd.mean=data2sdmean(data)
rules=c("Et->Luciferase","E2+siRIP140->RIP140","E2+siLCoR->LCoR","E2->0")
matp=read.rules(rules)
interval(sd.mean$mean,sd.mean$sd,matp,nrep=6)
}
