% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra.R
\name{mra}
\alias{mra}
\title{Modular Reponse Analysis function}
\usage{
mra(tab,matp,check=TRUE)
}
\arguments{
\item{tab}{A data.frame containing experimental data in a specific format (see details).}

\item{matp}{The perturbation matrix. Names of modules (rows) and perturbations (columns) must correspond to names of rows
and columns in tab.}

\item{check}{Logical. Should the dataset and perutbation matrix be checked for input errors?}
}
\value{
A list containing the connectivity map, the local responses matrix, the network responses matrix to perturbations
        and the basal line for all modules.
}
\description{
Calculation of connectivity coefficients between modules in biological networks
using modular response analysis.
}
\details{
It assumes that one perturbation must affect only one biological module of the network. This is specified as binary values in
the perturbation matrix. It also assumes that row names in data tables are the names of biological modules and column names are the
the names of perturbations.
}
\examples{
#It creates the connectivity map between 2 transcriptional nuclear coregulators
#(RIP140 and LCoR) and estrogen receptor alpha transcriptional activity reported by
#a luciferase gene. q-PCR data is stored in the package files used below in the function data.setup
#The model is obtained using the E2 stimulated condition.

data=data.setup(list(estr1_A,estr1_B,estr2_A,estr2_B,estr3_A,estr3_B))
sd.mean=data2sdmean(data)
rules=c("Et->Luciferase","E2+siRIP140->RIP140","E2+siLCoR->LCoR","E2->0")
matp=read.rules(rules)
mra(sd.mean$mean,matp,check=TRUE)
}
