% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.rules.R
\name{read.rules}
\alias{read.rules}
\title{Creates a perturbation matrix from a set of perturbation rules.}
\usage{
read.rules(obj)
}
\arguments{
\item{obj}{Either the path to a perturbation rules file without header or a vector of strings specifying the perturbation rules (see details).}
}
\description{
A function that takes a set of perturbation rules as input and creates a perturbation matrix to be used for
modular response analysis.
}
\details{
Perturbations rules are a set of strings specifying the action of perturbations upon modules in a network. In both cases modules
and perturbations must have the same names as the names of rows and columns in experimental data tables (names are case sensitive).

The rules syntax must be  ```Perturbation->Module``` for a single perturbation,```Perturbation->0``` to specify the basal line and
```0->Module``` to specify modules that were not perturbated but for which connectivity coming from other modules can be retrieved.
At least two rules and a basal line must be written.
}
\examples{
rules=c("Et->Luciferase","E2+siRIP140->RIP140","E2+siLCoR->LCoR","E2->0")
read.rules(rules)
}
