% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.ex.R
\docType{data}
\name{sd.ex}
\alias{sd.ex}
\title{standard deviation table from perturbations to the ERs,RARs, LCoR and RIP140 network}
\format{
A data frame with 8 rows (biological modules) and  15 variables (perturbations):
\describe{
   \item{Et}{Ethanol}
   \item{RA}{Retinoic Acid}
   \item{E2+RA}{Estradiol and retinoic acid}
   \item{siLCoR}{LCoR gene silencing RNA}
   \item{E2+siLCoR}{Retinoic acid and LCoR gene silencing RNA}
   \item{RA+siLCoR}{Retinoic acid and LCoR gene silencing RNA}
   \item{E2+RA+siLCoR}{Estradiol,retinoic acid and LCoR gene silencing RNA}
   \item{siRIP140}{RIP140 gene silencing RNA}
   \item{E2+siRIP140}{Estradiol and RIP140 gene silencing RNA}
   \item{RA+siRIP140}{Retinoic acid and RIP140 gene silencing RNA}
   \item{E2+RA+siRIP140}{Estradiol, retinoic acid and RIP140 gene silencing RNA}
   \item{siLCoR+siRIP140}{LCoR gene silencing RNA and RIP140 gene silencing RNA}
   \item{E2+siLCoR+siRIP140}{Estradiol, LCoR gene silencing RNA and RIP140 gene silencing RNA}
   \item{RA+siLCoR+siRIP140}{Retinoic acid, LCoR gene silencing RNA and RIP140 gene silencing RNA}
   \item{E2+RA+siLCoR+siRIP140}{Estradiol, retinoic acid, LCoR gene silencing RNA and RIP140 gene silencing RNA}
   ...
}
}
\source{
\url{EDF R\&D}
}
\usage{
sd.ex
}
\description{
standard deviation table from perturbations to the ERs,RARs, LCoR and RIP140 network
}
\keyword{datasets}
