% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.setup.R
\name{data.setup}
\alias{data.setup}
\title{Data setup for Modular Response Analysis}
\usage{
data.setup(obj, pert.tab = NULL, dec = ".")
}
\arguments{
\item{obj}{A character string specifying the path to a folder to be checked that contains the
experimental data tables and the perturbation rules file. It can be also several data tables
stored into a list or a single data.frames.}

\item{pert.tab}{Optional. The perturbations rules table to be checked for correct format.}

\item{dec}{The decimal separator as in ```base::read.csv```. If not "." (default) then usually ","}
}
\value{
If more than one data table is checked it returns a list of data.frames in the correct
format for the mra function. It also checks for correspondance with the perurbation rules file.
}
\description{
Checks for possible errors in data tables and in the perturbations matrix and returns the
correct format for input to the mra function.
}
\examples{
data=data.setup(list(estr1_A,estr1_B,estr2_A,estr2_B,estr3_A,estr3_B))
}
