% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2sdmean.R
\name{data2sdmean}
\alias{data2sdmean}
\title{Standard deviation and average calculation of a data set.}
\usage{
data2sdmean(x)
}
\arguments{
\item{x}{A list containing 2 or more tables.}
}
\value{
A list containing two matrices, the standard deviations and mean values for each variable in x.
}
\description{
This function computes the standard deviation and average values of a data set. Namely, if n replicates were generated,
then n data tables should be provided to ```data.setup``` and ```data2sdmean``` computes the average and standard deviation
of each cell over the n replicates.
}
\examples{
data=data.setup(list(estr1_A,estr1_B,estr2_A,estr2_B,estr3_A,estr3_B))
data2sdmean(data)
}
