% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra.R
\name{mra}
\alias{mra}
\title{Modular Reponse Analysis function}
\usage{
mra(tab,matp,check=TRUE)
}
\arguments{
\item{tab}{Data in the format returned by ```data.setup``` or a single data table in the format required by ```data.setup```.}

\item{matp}{A perturbation rule table, with rows corresponding to the MRA modules and columns to perturbations.}

\item{check}{Logical. If ```TRUE``` then data and perturbation rule are checked for consistency (by calling ```data.setup```).}
}
\value{
A list containing the connectivity map, the local responses matrix, the network responses matrix to perturbations
        and the basal line for all modules.
}
\description{
Calculation of connectivity coefficients between MRA modules
using modular response analysis.
}
\details{
A list containing the connectivity map, the local responses matrix, the network responses matrix to perturbations
and the basal line for all the modules.
}
\examples{
#It creates the connectivity map between 2 transcriptional nuclear coregulators
#(RIP140 and LCoR) and estrogen receptor alpha transcriptional activity reported by
#a luciferase gene. q-PCR data is stored in the package files used below in the function data.setup
#The model is obtained using the E2 stimulated condition.

data=data.setup(list(estr1_A,estr1_B,estr2_A,estr2_B,estr3_A,estr3_B))
sd.mean=data2sdmean(data)
rules=c("Et->Luciferase","E2+siRIP140->RIP140","E2+siLCoR->LCoR","E2->0")
matp=read.rules(rules)
mra(sd.mean$mean,matp,check=TRUE)
}
